feat: add voyage team client port and adapter, setup dtos
