+++ src/modules/voyage-team/application/dtos/response.dto.ts
@@ -1,0 +1,26 @@
+export interface GetCurrentVoyageTeamResponseDto extends VoyageTeamMember {}
+
+export interface VoyageTeamMember {
+  id: number;
+  voyageTeamId: number;
+  voyageTeam: VoyageTeam;
+  voyageRole: VoyageRole;
+}
+
+export interface VoyageStatus {
+  name: string;
+}
+
+interface Voyage {
+  status: VoyageStatus;
+}
+
+interface VoyageTeam {
+  name: string;
+  voyage: Voyage;
+  projectSubmitted: boolean;
+}
+
+export interface VoyageRole {
+  name: string;
+}
