+++ src/modules/voyage-team/ports/primary/voyageTeamClientPort.ts
@@ -1,0 +1,8 @@
+import { type GetUserRequestDto } from "@/modules/user/application/dtos/request.dtos";
+import { type GetCurrentVoyageTeamResponseDto } from "@/modules/voyage-team/application/dtos/response.dto";
+
+export interface VoyageTeamClientPort {
+  getCurrentVoyageTeam: (
+    user: GetUserRequestDto,
+  ) => GetCurrentVoyageTeamResponseDto;
+}
